JP-BE-PY-batch-2/class_22/user-permissions/requirements.txt

djangorestframework
djangorestframework-simplejwt